+++ CHANGELOG.md
@@ -4,6 +4,7 @@ https://keepachangelog.com/en/1.0.0/
 ## [1.31.2] - 2026-03-13
 - `TestMethodWithoutPublicModifier`: `internal` methods with xUnit no longer trigger this
 - `RecursiveOperatorOverload`: Fixed a false positive when a different overload is selected
+- `SwitchDoesNotHandleAllEnumOptions`: Correctly simplifies when `global using` statements are present
 
 ## [1.31.1] - 2026-01-25
 - `MultipleFromBodyParameters`: Support Minimal Web APIs (lambda expressions)